feat: [issue-13] agregando carga de datos desde kaggle
