--- scripts_py/old/main_kaggle_log.py
@@ -1,1 +1,0 @@
-#run load_kaggle_data.py
