--- scripts_py/old/normalization.py
@@ -1,5 +1,0 @@
-"""### Normalización"""
-
-df_limpio = formatear_fecha(df_filtrado)
-df_limpio.write_csv("log_limpio.csv")
-df_limpio
